Abstracting outbound side of transport (#13293)
* Abstracting outbound side of transport


* Making outbound handler protocol dependent via inbound handler


* Fixing precommit


* Addressing comments


* Fixing precommit


* Cleaning up code


* Addressing comments


* Cleaning up


* Addressing comments


* Abstracting InboundHandlerTests


* Abstracting TransportLoggerTests


---------
