+++ server/src/main/java/org/opensearch/transport/nativeprotocol/NetworkMessage.java
@@ -29,11 +29,12 @@
  * GitHub history for details.
  */
 
-package org.opensearch.transport;
+package org.opensearch.transport.nativeprotocol;
 
 import org.opensearch.Version;
 import org.opensearch.common.util.concurrent.ThreadContext;
 import org.opensearch.core.common.io.stream.Writeable;
+import org.opensearch.transport.TransportStatus;
 
 /**
  * Represents a transport message sent over the network. Subclasses implement serialization and
