+++ server/src/test/java/org/opensearch/transport/InboundPipelineTests.java
@@ -52,7 +52,6 @@ import org.opensearch.test.OpenSearchTestCase;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -63,12 +62,22 @@ import java.util.function.Supplier;
 
 import static org.hamcrest.Matchers.instanceOf;
 
-public class InboundPipelineTests extends OpenSearchTestCase {
+public abstract class InboundPipelineTests extends OpenSearchTestCase {
 
     private static final int BYTE_THRESHOLD = 128 * 1024;
-    private final ThreadContext threadContext = new ThreadContext(Settings.EMPTY);
+    public final ThreadContext threadContext = new ThreadContext(Settings.EMPTY);
 
-    public void testPipelineHandling() throws IOException {
+    protected abstract BytesReference serialize(
+        boolean isRequest,
+        Version version,
+        boolean handshake,
+        boolean compress,
+        String action,
+        long requestId,
+        String value
+    ) throws IOException;
+
+    public void testPipelineHandlingForNativeProtocol() throws IOException {
         final List<Tuple<MessageData, Exception>> expected = new ArrayList<>();
         final List<Tuple<MessageData, Exception>> actual = new ArrayList<>();
         final List<ReleasableBytesReference> toRelease = new ArrayList<>();
@@ -85,10 +94,10 @@ public class InboundPipelineTests extends OpenSearchTestCase {
                     actualData = new MessageData(version, requestId, isRequest, isCompressed, header.getActionName(), null);
                 } else if (isRequest) {
                     final TestRequest request = new TestRequest(message.openOrGetStreamInput());
-                    actualData = new MessageData(version, requestId, isRequest, isCompressed, header.getActionName(), request.value);
+                    actualData = new MessageData(version, requestId, isRequest, isCompressed, header.getActionName(), request.getValue());
                 } else {
                     final TestResponse response = new TestResponse(message.openOrGetStreamInput());
-                    actualData = new MessageData(version, requestId, isRequest, isCompressed, null, response.value);
+                    actualData = new MessageData(version, requestId, isRequest, isCompressed, null, response.getValue());
                 }
                 actual.add(new Tuple<>(actualData, message.getException()));
             } catch (IOException e) {
@@ -127,49 +136,23 @@ public class InboundPipelineTests extends OpenSearchTestCase {
                     final MessageData messageData;
                     Exception expectedExceptionClass = null;
 
-                    OutboundMessage message;
+                    // NativeOutboundMessage message;
+                    final BytesReference reference;
                     if (isRequest) {
                         if (rarely()) {
                             messageData = new MessageData(version, requestId, true, isCompressed, breakThisAction, null);
-                            message = new OutboundMessage.Request(
-                                threadContext,
-                                new String[0],
-                                new TestRequest(value),
-                                version,
-                                breakThisAction,
-                                requestId,
-                                false,
-                                isCompressed
-                            );
+                            reference = serialize(true, version, false, isCompressed, breakThisAction, requestId, value);
                             expectedExceptionClass = new CircuitBreakingException("", CircuitBreaker.Durability.PERMANENT);
                         } else {
                             messageData = new MessageData(version, requestId, true, isCompressed, actionName, value);
-                            message = new OutboundMessage.Request(
-                                threadContext,
-                                new String[0],
-                                new TestRequest(value),
-                                version,
-                                actionName,
-                                requestId,
-                                false,
-                                isCompressed
-                            );
+                            reference = serialize(true, version, false, isCompressed, actionName, requestId, value);
                         }
                     } else {
                         messageData = new MessageData(version, requestId, false, isCompressed, null, value);
-                        message = new OutboundMessage.Response(
-                            threadContext,
-                            Collections.emptySet(),
-                            new TestResponse(value),
-                            version,
-                            requestId,
-                            false,
-                            isCompressed
-                        );
+                        reference = serialize(false, version, false, isCompressed, actionName, requestId, value);
                     }
 
                     expected.add(new Tuple<>(messageData, expectedExceptionClass));
-                    final BytesReference reference = message.serialize(new BytesStreamOutput());
                     Streams.copy(reference.streamInput(), streamOutput);
                 }
 
@@ -230,31 +213,7 @@ public class InboundPipelineTests extends OpenSearchTestCase {
             final boolean isRequest = randomBoolean();
             final long requestId = randomNonNegativeLong();
 
-            OutboundMessage message;
-            if (isRequest) {
-                message = new OutboundMessage.Request(
-                    threadContext,
-                    new String[0],
-                    new TestRequest(value),
-                    invalidVersion,
-                    actionName,
-                    requestId,
-                    false,
-                    false
-                );
-            } else {
-                message = new OutboundMessage.Response(
-                    threadContext,
-                    Collections.emptySet(),
-                    new TestResponse(value),
-                    invalidVersion,
-                    requestId,
-                    false,
-                    false
-                );
-            }
-
-            final BytesReference reference = message.serialize(streamOutput);
+            final BytesReference reference = serialize(isRequest, invalidVersion, false, false, actionName, requestId, value);
             try (ReleasableBytesReference releasable = ReleasableBytesReference.wrap(reference)) {
                 expectThrows(IllegalStateException.class, () -> pipeline.handleBytes(new FakeTcpChannel(), releasable));
             }
@@ -284,31 +243,7 @@ public class InboundPipelineTests extends OpenSearchTestCase {
             final boolean isRequest = randomBoolean();
             final long requestId = randomNonNegativeLong();
 
-            OutboundMessage message;
-            if (isRequest) {
-                message = new OutboundMessage.Request(
-                    threadContext,
-                    new String[0],
-                    new TestRequest(value),
-                    version,
-                    actionName,
-                    requestId,
-                    false,
-                    false
-                );
-            } else {
-                message = new OutboundMessage.Response(
-                    threadContext,
-                    Collections.emptySet(),
-                    new TestResponse(value),
-                    version,
-                    requestId,
-                    false,
-                    false
-                );
-            }
-
-            final BytesReference reference = message.serialize(streamOutput);
+            final BytesReference reference = serialize(isRequest, version, false, false, actionName, requestId, value);
             final int fixedHeaderSize = TcpHeader.headerSize(Version.CURRENT);
             final int variableHeaderSize = reference.getInt(fixedHeaderSize - 4);
             final int totalHeaderSize = fixedHeaderSize + variableHeaderSize;
