+++ test/framework/src/main/java/org/opensearch/transport/TestResponse.java
@@ -54,4 +54,8 @@ public class TestResponse extends TransportResponse {
     public void writeTo(StreamOutput out) throws IOException {
         out.writeString(value);
     }
+
+    public String getValue() {
+        return value;
+    }
 }
